refactor: share the ThreadSafeMessageBox mock between gtests
Move MockUIInterface out of test_deprecation.cpp into gtest/utils and add
ConnectMockUIInterface / DisconnectMockUIInterface helpers there, so other gtests
that exercise code which notifies the user via uiInterface.ThreadSafeMessageBox can
reuse them. The gtest harness connects no slot to that signal, so without a mock its
signals2 combiner throws on zero slots; DisconnectMockUIInterface also clears any
shutdown the test requested.

No behaviour change: test_deprecation.cpp now connects and disconnects the shared
mock via these helpers instead of its own local copy.
